[zep noup][tf-m] bootutil: Parse key ID for built-in keys
When MCUBOOT_BUILTIN_KEY is enabled, the key id TLV entry is added
to the image with the new IMAGE_TLV_KEYID type. Parse this entry while
validating the image to identify the key used to sign the image.
In this scenario the IMAGE_TLV_KEYHASH or IMAGE_TLV_PUBKEY entries
are not needed.

This replaces the previous PoC code that simply assigned the
image_index to the key ID.

+++ boot/bootutil/src/bootutil_priv.h
@@ -3,7 +3,7 @@
  *
  * Copyright (c) 2017-2020 Linaro LTD
  * Copyright (c) 2017-2019 JUUL Labs
- * Copyright (c) 2019-2021 Arm Limited
+ * Copyright (c) 2019-2025 Arm Limited
  *
  * Original license:
  *
@@ -288,7 +288,7 @@ struct boot_sector_buffer {
  * an image and mlen to length of the hash.
  */
 fih_ret bootutil_verify_sig(uint8_t *msg, uint32_t mlen, uint8_t *sig,
-                            size_t slen, uint8_t key_id);
+                            size_t slen, uint32_t key_id);
 
 fih_ret boot_fih_memequal(const void *s1, const void *s2, size_t n);
